Create Individual assignment - Image Analysis
This is a compilation of the two individual assignments made for the course in Computer Vision from Tilburg University. For these assignments, the goal was to analyze, enhance images as well as identify objects in images using a collection of methods in Python. High marks were received for both of them.