Add files via upload
This is an overview of two individual assignments for the course in Computer Vision called Image Analysis at Tilburg University. Both assignments has as goal to filter, enhance images and the course also taught a wide range of methods to do these things as well as segment objects in images.